Renames explorerNodes
